close picture refactored
