+++ js/util.js
@@ -7,33 +7,6 @@ const getRandomValue = (a, b) => {
 
 const isEscapeKey = (evt) => evt.key === 'Escape';
 
-const createClosePictureHandle = (pictureContainer, closeButton) => {
-  const hidePictureContainer = () => {
-    pictureContainer.classList.add('hidden');
-    document.body.classList.remove('modal-open');
-  };
-
-  const onEscapeClick = (evt) => {
-    if (isEscapeKey(evt)) {
-      evt.preventDefault();
-      document.removeEventListener('keydown', onEscapeClick);
-      hidePictureContainer();
-    }
-  };
-
-  closeButton.addEventListener('click', () => {
-    document.removeEventListener('keydown', onEscapeClick);
-    hidePictureContainer();
-  });
-
-  return (evt) => {
-    evt.preventDefault();
-    document.body.classList.add('modal-open');
-    pictureContainer.classList.remove('hidden');
-    document.addEventListener('keydown', onEscapeClick);
-  };
-};
-
 const isFocusedElement = (element) => {
   element.addEventListener('keydown', (evt) => {
     if (isEscapeKey(evt)) {
@@ -42,4 +15,4 @@ const isFocusedElement = (element) => {
   });
 };
 
-export {getRandomValue, isEscapeKey, createClosePictureHandle, isFocusedElement};
+export {getRandomValue, isEscapeKey, isFocusedElement};
